Started stubbing out REST endpoints with new arch.
